requirements/base.txt
Django==5.0.1
psycopg2-binary==2.9.9
djangorestframework==3.14.0
pandas==2.2.0
drf-yasg ==1.21.7
serpy==0.3.1
cerberus==1.3.5
